Move Logr::Logger class into its own file
